+++ scanpipe/tests/test_models.py
@@ -62,7 +62,6 @@ from scanpipe.models import get_project_work_directory
 from scanpipe.models import posix_regex_to_django_regex_lookup
 from scanpipe.pipes.fetch import Download
 from scanpipe.pipes.input import copy_input
-from scanpipe.pipes.input import copy_inputs
 from scanpipe.tests import dependency_data1
 from scanpipe.tests import dependency_data2
 from scanpipe.tests import license_policies_index
@@ -369,9 +368,32 @@ class ScanPipeModelsTest(TestCase):
         self.project1.refresh_from_db()
         self.assertEqual({"filename": "source"}, self.project1.input_sources)
 
+    def test_scanpipe_project_model_delete_input(self):
+        self.assertEqual({}, self.project1.input_sources)
+        self.assertEqual([], list(self.project1.inputs()))
+        deleted = self.project1.delete_input(name="not_existing")
+        self.assertFalse(deleted)
+
+        file_location = self.data_location / "notice.NOTICE"
+        copy_input(file_location, self.project1.input_path)
+        self.project1.add_input_source(
+            filename=file_location.name, source="uploaded", save=True
+        )
+        self.project1.refresh_from_db()
+        self.assertEqual({file_location.name: "uploaded"}, self.project1.input_sources)
+        self.assertEqual(
+            [file_location.name], [path.name for path in self.project1.inputs()]
+        )
+
+        deleted = self.project1.delete_input(name=file_location.name)
+        self.assertTrue(deleted)
+        self.project1.refresh_from_db()
+        self.assertEqual({}, self.project1.input_sources)
+        self.assertEqual([], list(self.project1.inputs()))
+
     def test_scanpipe_project_model_add_downloads(self):
         file_location = self.data_location / "notice.NOTICE"
-        copy_inputs([file_location], self.project1.tmp_path)
+        copy_input(file_location, self.project1.tmp_path)
 
         download = Download(
             uri="https://example.com/filename.zip",